Added styling to title section
