+++ src/components/common-title/common-title.styles.scss
@@ -1,6 +1,10 @@
 .title-link {
+    display: flex;
+    align-items: center;
     color: #d5a10f;
     text-decoration: none;
     font-size: 1rem;
     font-weight: 400;
+    height: 50%;
+    // border: 1px solid red;
 }
