+++ src/components/title-section/title-section.component.jsx
@@ -1,14 +1,19 @@
-import React from 'react'
+import React from 'react';
+
+import './title-section.styles.scss'
+
 import CommonTitle from '../common-title/common-title.component';
 
 const TitleSection = () => {
     return (
-        <div>
-            <CommonTitle>New Article</CommonTitle>
-            <CommonTitle>View Articles</CommonTitle>
+        <div className='title-section-container'>
+            <div className='titles-container'>
+                <CommonTitle>New Article</CommonTitle>
+                <CommonTitle>View Articles</CommonTitle>
+            </div>
             <hr />
         </div>
-    )
+    );
 }
 
 export default TitleSection
