+++ src/components/title-section/title-section.styles.scss
@@ -1,0 +1,16 @@
+.title-section-container {
+    display: flex;
+    flex-direction: column;
+    justify-content: space-evenly;
+    // border: 1px solid red;
+}
+
+.titles-container {
+    display: grid;
+    width: 100%;
+    grid-template-columns: 1fr 1fr;
+}
+
+hr {
+    border: 0.1px solid #d5a10f;
+}
