+++ src/pages/articles-page/articles-page.component.jsx
@@ -1,5 +1,6 @@
 import React from 'react';
 
+
 const ArticlesPage = () => {
     return (
         <div>
