--- src/pages/create-article-page.component.jsx
@@ -1,16 +1,0 @@
-import React from 'react';
-import TitleSection from '../components/title-section/title-section.component';
-import ImageSection from '../components/image-section/image-section.component';
-import FormSection from '../components/form-section/form-section.component';
-
-const CreateArticlePage = () => {
-    return (
-        <div>
-            <TitleSection />
-            <ImageSection />
-            <FormSection />
-        </div>
-    )
-}
-
-export default CreateArticlePage
