+++ src/pages/create-article-page/create-article-page.styles.scss
@@ -1,0 +1,15 @@
+.create-article-page-container {
+    border: 1px solid yellow;
+    display: flex;
+    justify-content: center;
+    min-height: 100vh;
+}
+
+.create-article-page-wrapper {
+    display: grid;
+    grid-template-rows: 0.3fr 1fr 1.5fr;
+    border: 1px solid blue;
+    width: 80%;
+    padding-inline: 1.5rem;
+    margin-block: 5%;
+}
